refactor: introduce shared types package for dependency inversion
Move MediaDownloadRequest to internal/types so both client and commands
can reference the same type without circular dependencies. This enables
true dependency injection where interfaces don't depend on concrete
implementation packages.

Changes:
- Create internal/types/media.go with MediaDownloadRequest
- Update client package to use types.MediaDownloadRequest
- Update commands interfaces to use types.MediaDownloadRequest
- Add StartSync method to WAClient interface
- Rename whatsmeow types import to waTypes to avoid collision
