+++ internal/client/client.go
@@ -12,10 +12,11 @@ import (
 
 	_ "github.com/mattn/go-sqlite3"
 	"github.com/mdp/qrterminal"
+	"github.com/vicentereig/whatsapp-cli/internal/types"
 	"go.mau.fi/whatsmeow"
 	waProto "go.mau.fi/whatsmeow/binary/proto"
 	"go.mau.fi/whatsmeow/store/sqlstore"
-	"go.mau.fi/whatsmeow/types"
+	waTypes "go.mau.fi/whatsmeow/types"
 	"go.mau.fi/whatsmeow/types/events"
 	waLog "go.mau.fi/whatsmeow/util/log"
 	"google.golang.org/protobuf/proto"
@@ -25,8 +26,8 @@ type WAClient struct {
 	client          *whatsmeow.Client
 	storeDir        string
 	eventHandler    func(interface{})
-	contactLookup   func(ctx context.Context, user types.JID) (types.ContactInfo, error)
-	groupInfoLookup func(ctx context.Context, jid types.JID) (*types.GroupInfo, error)
+	contactLookup   func(ctx context.Context, user waTypes.JID) (waTypes.ContactInfo, error)
+	groupInfoLookup func(ctx context.Context, jid waTypes.JID) (*waTypes.GroupInfo, error)
 }
 
 type MediaInfo struct {
@@ -52,15 +53,6 @@ type MessageDetails struct {
 	Media     *MediaInfo
 }
 
-type MediaDownloadRequest struct {
-	DirectPath    string
-	MediaKey      []byte
-	FileSHA256    []byte
-	FileEncSHA256 []byte
-	FileLength    uint64
-	MediaType     string
-	MimeType      string
-}
 
 func NewWAClient(storeDir string) (*WAClient, error) {
 	// Create store directory
@@ -167,26 +159,26 @@ func (w *WAClient) AddEventHandler(handler func(interface{})) {
 	w.client.AddEventHandler(handler)
 }
 
-func contactLookupFunc(cli *whatsmeow.Client) func(ctx context.Context, user types.JID) (types.ContactInfo, error) {
+func contactLookupFunc(cli *whatsmeow.Client) func(ctx context.Context, user waTypes.JID) (waTypes.ContactInfo, error) {
 	if cli == nil || cli.Store == nil || cli.Store.Contacts == nil {
 		return nil
 	}
-	return func(ctx context.Context, user types.JID) (types.ContactInfo, error) {
+	return func(ctx context.Context, user waTypes.JID) (waTypes.ContactInfo, error) {
 		return cli.Store.Contacts.GetContact(ctx, user)
 	}
 }
 
-func groupInfoLookupFunc(cli *whatsmeow.Client) func(ctx context.Context, jid types.JID) (*types.GroupInfo, error) {
+func groupInfoLookupFunc(cli *whatsmeow.Client) func(ctx context.Context, jid waTypes.JID) (*waTypes.GroupInfo, error) {
 	if cli == nil {
 		return nil
 	}
-	return func(ctx context.Context, jid types.JID) (*types.GroupInfo, error) {
+	return func(ctx context.Context, jid waTypes.JID) (*waTypes.GroupInfo, error) {
 		info, err := cli.GetGroupInfo(ctx, jid)
 		return info, err
 	}
 }
 
-func bestContactName(info types.ContactInfo) string {
+func bestContactName(info waTypes.ContactInfo) string {
 	if !info.Found {
 		return ""
 	}
@@ -214,10 +206,10 @@ func (w *WAClient) ResolveChatName(ctx context.Context, chatJID string, msg *eve
 	}
 	fallback := chatJID
 
-	parsed, err := types.ParseJID(chatJID)
+	parsed, err := waTypes.ParseJID(chatJID)
 	if err == nil {
 		// Group chats
-		if parsed.Server == types.GroupServer || parsed.IsBroadcastList() {
+		if parsed.Server == waTypes.GroupServer || parsed.IsBroadcastList() {
 			if w.groupInfoLookup != nil {
 				if info, err := w.groupInfoLookup(ctx, parsed); err == nil && info != nil {
 					if name := strings.TrimSpace(info.GroupName.Name); name != "" {
@@ -258,14 +250,14 @@ func (w *WAClient) StartSync(ctx context.Context, eventHandler func(interface{})
 	return nil
 }
 
-func parseJID(recipient string) (types.JID, error) {
+func parseJID(recipient string) (waTypes.JID, error) {
 	// If already a JID, parse it
 	if contains(recipient, "@") {
-		return types.ParseJID(recipient)
+		return waTypes.ParseJID(recipient)
 	}
 
 	// Otherwise, assume it's a phone number
-	return types.JID{
+	return waTypes.JID{
 		User:   recipient,
 		Server: "s.whatsapp.net",
 	}, nil
@@ -280,7 +272,7 @@ func contains(s, substr string) bool {
 	return false
 }
 
-func (w *WAClient) DownloadMediaToFile(ctx context.Context, req MediaDownloadRequest, targetPath string) (int64, error) {
+func (w *WAClient) DownloadMediaToFile(ctx context.Context, req types.MediaDownloadRequest, targetPath string) (int64, error) {
 	if w == nil || w.client == nil {
 		return 0, fmt.Errorf("whatsapp client is not initialized")
 	}
